+++ Assets/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.5.0] - 2023-12-22
+
+### Added
+
+- Added ledge push force to Kinematic Character Motor
+  We noticed that the character could somewhat float off ledges
+  So we thought it would be nice to add ability for motor to push 
+  the character off the ledge if they are past the ledge stability distance
+
 ## [3.4.4] - 2023-12-17
 
 ### Added